chore(rename): change lib tests to src
