+++ packages/holocron-dev-server/__tests__/src/components/RegisterModule.spec.jsx
