+++ packages/holocron-dev-server/__tests__/src/utils/module-map.spec.js
