+++ packages/holocron-dev-server/__tests__/src/utils/statics.spec.js
