+++ packages/holocron-dev-server/__tests__/src/webpack/configs/fragments.spec.js
@@ -30,8 +30,8 @@ import {
   createEnvironmentDefinitionsConfigFragment,
   createResolverConfigFragment,
   createWatchOptionsConfigFragment,
-} from '../../../../src/webpack/configs';
-import { getWebpackVersion } from '../../../../src/webpack';
+} from '../../../../src/webpack/configs/fragments';
+import { getWebpackVersion } from '../../../../src/webpack/helpers';
 
 jest.mock('../../../../src/webpack/helpers', () => {
   const originalModule = jest.requireActual('../../../../src/webpack/helpers');
